requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-haystack==2.4.0
django-storages==1.1.8
django-tables2==1.0.4
elasticsearch==1.6.0
gunicorn==19.3.0
Pillow==2.8.1
psycopg2==2.6.1
simplejson==3.7.3
six==1.9.0
static3==0.6.1
urllib3==1.10.4
whitenoise==1.0.6
